FS Utils: Move all utils to a common lib
