+++ common/lib/is-errno-exception.ts
