+++ src/storage/user-data.ts
@@ -5,8 +5,8 @@ import { SupportedPHPVersion, SupportedPHPVersions } from '@php-wasm/universal';
 import * as Sentry from '@sentry/electron/main';
 import { readFile, writeFile } from 'atomically';
 import { LOCKFILE_STALE_TIME, LOCKFILE_WAIT_TIME } from 'common/constants';
+import { isErrnoException } from 'common/lib/is-errno-exception';
 import { lockFileAsync, unlockFileAsync } from 'common/lib/lockfile';
-import { isErrnoException } from 'src/lib/is-errno-exception';
 import { sanitizeUnstructuredData, sanitizeUserpath } from 'src/lib/sanitize-for-logging';
 import { sortSites } from 'src/lib/sort-sites';
 import { getUserDataFilePath, getUserDataLockFilePath } from 'src/storage/paths';
